Fix types of JSON fields
The HTTP status code int can exceed 255, e.g. for 404.
“total” doesn’t need to be signed because it’s a count. There will not be a negative count.

+++ src/words.rs
@@ -11,8 +11,8 @@ use self::spinner::{SpinnerBuilder};
 
 #[derive(Serialize, Deserialize)]
 struct Entry {
-    status: u8,
-    total: i64,
+    status: u32,
+    total: u64,
     results: Vec<Result>,
 }
